requirements.txt
google.generativeai
python-dotenv
requests
PyPDF2
SpeechRecognition
docx
flask
python-docx 
